# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: 上岸 V1 API
  description: iOS 客户端使用的只读与学习业务接口；所有业务路由统一使用 /api/v1 前缀。
  version: 0.1.0-rc1
servers:
- url: http://localhost
  description: Generated server url
paths:
...
  /api/v1/lessons/{lessonId}:
    get:
      tags:
      - catalog-controller
      operationId: lesson
      parameters:
      - name: lessonId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LessonResponse"
      security:
      - bearerAuth: []
  /api/v1/lessons/{lessonId}/study-content:
    get:
      tags:
      - catalog-controller
      operationId: studyContent
      parameters:
      - name: lessonId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LessonStudyContentResponse"
      security:
      - bearerAuth: []
...
  /api/v1/courses:
    get:
      tags:
      - catalog-controller
      operationId: courses
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CourseResponse"
      security:
      - bearerAuth: []
  /api/v1/courses/{courseId}:
    get:
      tags:
      - catalog-controller
      operationId: course
      parameters:
      - name: courseId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CourseDetailResponse"
      security:
      - bearerAuth: []
components:
  schemas:
...
    LessonResponse:
      type: object
      properties:
        id:
          type: string
        courseId:
          type: string
        title:
          type: string
        durationMs:
          type: integer
          format: int64
        sortOrder:
          type: integer
          format: int32
    LessonStudyContentResponse:
      type: object
      properties:
        lessonId:
          type: string
        fullText:
          type: string
        summaryMarkdown:
          type: string
        updatedAt:
          type: string
          format: date-time
...
    CourseResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
    CourseDetailResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/LessonResponse"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT